Refactor: math_kernel_op split op files re-order ops
