+++ python/pyabacus/CONTRIBUTING.md
@@ -191,7 +191,8 @@ list(APPEND _diago
     ${HSOLVER_PATH}/diago_iter_assist.cpp
     ${HSOLVER_PATH}/kernels/dngvd_op.cpp
     ${BASE_PATH}/kernels/math_kernel_op.cpp
-    ${BASE_PATH}/kernels/math_op.cpp
+    ${BASE_PATH}/kernels/math_kernel_op_vec.cpp
+    ${BASE_PATH}/kernels/math_ylm_op.cpp
     ${BASE_PATH}/module_device/device.cpp
     ${BASE_PATH}/module_device/memory_op.cpp
     ${PSI_PATH}/psi.cpp
