+++ source/module_hsolver/diago_bpcg.h
@@ -349,8 +349,8 @@ class DiagoBPCG
     // note: these operators use template parameter base_device::Device_*
     // defined in module_base/module_device/types.h
     // different from ct_Device!
-    using calc_grad_with_block_op = ModuleBase::calc_grad_with_block_op<T, Device>;
-    using line_minimize_with_block_op = ModuleBase::line_minimize_with_block_op<T, Device>;
+    using calc_grad_with_block_op = calc_grad_with_block_op<T, Device>;
+    using line_minimize_with_block_op = line_minimize_with_block_op<T, Device>;
     using gemm_op = ModuleBase::gemm_op<T, Device>;
 
 };
