+++ source/module_hsolver/diago_david.cpp
@@ -966,7 +966,7 @@ void DiagoDavid<T, Device>::SchmidtOrth(const int& dim,
     else
     {
         // psi_m = psi_m / psi_norm
-        ModuleBase::vector_div_constant_op<T, Device>()(dim, psi_m, psi_m, psi_norm);
+        ModuleBase::vector_mul_real_op<T, Device>()(dim, psi_m, psi_m, 1.0 / psi_norm);
         // for (int i = 0; i < npw; i++)
         // {
         //     psi_m[i] /= psi_norm;
